Migrated from old Args to new Args structure
