+++ src/auto_flutter_dig/model/argument/__init__.py
@@ -1,2 +1,1 @@
-from .argument import Arg
 from .arguments import Args
