+++ src/auto_flutter_dig/task/_list.py
@@ -10,7 +10,6 @@ from .flutter.build.stub import FlutterBuildStub
 from .flutter.doctor import FlutterDoctor
 from .flutter.generator import FlutterGeneratorTask
 from .flutter.pub.get import FlutterPubGet
-from .options import ParseOptions
 from .project.init import ProjectInit
 from .project.read import ProjectRead
 from .project.save import ProjectSave
@@ -22,7 +21,6 @@ task_list: Dict[TaskId, TaskIdentity] = dict(
             Help.identity.to_map(),
             SetupEdit.identity.to_map(),
             Setup.identity.to_map(),
-            ParseOptions.identity.to_map(),
             ProjectRead.identity.to_map(),
             ProjectRead.identity_skip.to_map(),
             ProjectInit.identity.to_map(),
