+++ src/auto_flutter_dig/task/config/flavor.py
@@ -26,7 +26,7 @@ class ConfigFlavor(_BaseConfigTask):
 
         has_change: bool = False
 
-        add_flavor = args.get_value(self.option_add)
+        add_flavor = args.get(self.option_add)
         if not add_flavor is None and len(add_flavor) > 0:
             self._print("    Adding flavor " + add_flavor)
             if add_flavor in project.flavors:
@@ -39,7 +39,7 @@ class ConfigFlavor(_BaseConfigTask):
             project.flavors.append(add_flavor)
             has_change = True
 
-        rem_flavor = args.get_value(self.option_remove)
+        rem_flavor = args.get(self.option_remove)
         if not rem_flavor is None and len(rem_flavor) > 0:
             self._print("    Removing flavor " + rem_flavor)
             if not rem_flavor in project.flavors:
@@ -56,8 +56,8 @@ class ConfigFlavor(_BaseConfigTask):
                     if not config.flavored is None and rem_flavor in config.flavored:
                         config.flavored.pop(rem_flavor)
 
-        ren_flavor = args.get_value(self.option_rename)
-        to_flavor = args.get_value(self.option_toname)
+        ren_flavor = args.get(self.option_rename)
+        to_flavor = args.get(self.option_toname)
         has_ren = not ren_flavor is None and len(ren_flavor) > 0
         has_to = not to_flavor is None and len(to_flavor) > 0
         if has_ren or has_to:
