+++ src/auto_flutter_dig/task/firebase/upload.py
@@ -41,7 +41,7 @@ class FirebaseBuildUpload(BaseProcessTask):
         ]
 
     def _create_process(self, args: Args) -> ProcessOrResult:
-        filename = args.get_value("output")
+        filename = args.get("output")
         if filename is None or len(filename) <= 0:
             return TaskResult(
                 args, AssertionError("Previous task does not have output")
@@ -54,7 +54,7 @@ class FirebaseBuildUpload(BaseProcessTask):
             )
 
         file = file.absolute()
-        google_id = args.get_value(FirebaseBuildValidate.ARG_FIREBASE_GOOGLE_ID)
+        google_id = args.get(FirebaseBuildValidate.ARG_FIREBASE_GOOGLE_ID)
         if google_id is None or len(google_id) <= 0:
             return TaskResult(args, AssertionError("Google app id not found"))
 
@@ -67,19 +67,19 @@ class FirebaseBuildUpload(BaseProcessTask):
         ]
 
         _If.not_none(
-            args.get_value(self.__options["notes"]),
+            args.get(self.__options["notes"]),
             lambda notes: arguments.extend(("--release-notes", notes)),
             lambda: None,
         )
 
         _If.not_none(
-            args.get_value(self.__options["testers"]),
+            args.get(self.__options["testers"]),
             lambda testers: arguments.extend(("--testers", testers)),
             lambda: None,
         )
 
         _If.not_none(
-            args.get_value(self.__options["groups"]),
+            args.get(self.__options["groups"]),
             lambda groups: arguments.extend(("--groups", groups)),
             lambda: None,
         )
