+++ src/auto_flutter_dig/task/flutter/exec.py
@@ -4,10 +4,8 @@ from ...core.config import Config
 from ...core.process.process import Process
 from ...core.string import SB
 from ...model.argument.option import OptionAll
-from ...model.argument.option.opt_all_codec import ArgsDecode
 from ...model.task import *
 from ...task.identity import FlutterTaskIdentity
-from ..options import ParseOptions
 from ..project.read import ProjectRead
 from ._const import FLUTTER_CONFIG_KEY_PATH, FLUTTER_DISABLE_VERSION_CHECK
 
@@ -39,8 +37,8 @@ class Flutter(Task):
 
     def require(self) -> List[TaskId]:
         if self._project:
-            return [ParseOptions.identity.id, ProjectRead.identity.id]
-        return [ParseOptions.identity.id, ProjectRead.identity_skip.id]
+            return [ProjectRead.identity.id]
+        return [ProjectRead.identity_skip.id]
 
     def execute(self, args: Args) -> TaskResult:
         flutter = Config.get_path(FLUTTER_CONFIG_KEY_PATH)
@@ -55,11 +53,11 @@ class Flutter(Task):
 
         if not self._command is None and len(self._command) > 0:
             if self._command_args:
-                self._command.extend(ArgsDecode(args).all())
+                self._command.extend(args.get_all(OptionAll()))
             p = Process.create(flutter, arguments=self._command, writer=writer)
         else:
             arguments = [FLUTTER_DISABLE_VERSION_CHECK]
-            arguments.extend(ArgsDecode(args).all())
+            arguments.extend(args.get_all(OptionAll()))
             p = Process.create(flutter, arguments=arguments, writer=writer)
         output = p.try_run()
 
@@ -67,7 +65,7 @@ class Flutter(Task):
             self._print(p.output)
 
         if self._output_arg:
-            args.add_arg("output", p.output)
+            args.add("output", p.output)
 
         if isinstance(output, BaseException):
             return TaskResult(args, error=output)
