+++ src/auto_flutter_dig/task/options.py
@@ -1,136 +1,239 @@
-from getopt import GetoptError, gnu_getopt
-from sys import argv as sys_argv
-from typing import List, Optional
+from __future__ import annotations
 
-from ..core.session import Session
-from ..core.utils import _Dict, _Iterable
-from ..model.argument import Arg
-from ..model.argument.option import (
-    LongOption,
-    LongShortOptionWithValue,
-    Option,
-    OptionAll,
-    OptionWithValue,
-    ShortOption,
+from sys import argv as sys_argv
+from typing import Dict, Generic, Optional, Type, TypeVar, Union
+
+from ..model.argument.option import *
+from ..model.argument.option.error import (
+    OptionInvalidFormat,
+    OptionNotFound,
+    OptionRequireValue,
 )
-from ..model.argument.option.opt_all_codec import ArgsEncode
 from ..model.task import *
-from .help_stub import HelpStub
-from .identity import AflutterTaskIdentity
+
+Argument = str
+Group = str
+GroupedOptions = Dict[Group, Option]
+OptionsByArgument = Dict[Argument, GroupedOptions]
+
+T = TypeVar("T", bound=Option)
+
+
+class _Helper(Generic[T]):
+    def __init__(
+        self, option: T, group: Union[Group, TaskIdentity], cls: Type[T]
+    ) -> None:
+        self.option: T = option
+        self.group: Group = ""
+        if isinstance(group, Group):
+            self.group = group
+        elif isinstance(group, TaskIdentity):
+            self.group = group.group
+
+        self.has_value: bool = isinstance(option, OptionWithValue)
+        self.argument: Argument = ""
+        if cls is LongOption:
+            assert isinstance(option, LongOption)
+            self.argument = option.long
+        elif cls is ShortOption:
+            assert isinstance(option, ShortOption)
+            self.argument = option.short
+        elif cls is PositionalOption:
+            assert isinstance(option, PositionalOption)
+            self.argument = str(option.position)
+        pass
+
+    def into(self, target: Dict[Argument, Dict[Group, _Helper[T]]]):
+        if not self.argument in target:
+            target[self.argument] = {}
+        target[self.argument][self.group] = self
+
+
+class _ShortOptionMaybeWithValue(ShortOptionWithValue):
+    ...
+
+
+class _LongOptionMaybeWithValue(LongOptionWithValue):
+    ...
 
 
 class ParseOptions(Task):
-    option_stack_trace = LongOption(
-        "aflutter-stack-trace", "Show stacktrace of task output"
-    )
-    __options = {
-        "stack-trace": LongOption(
-            "aflutter-stack-trace", "Show stack trace of task output error"
-        ),
-        "help": LongShortOptionWithValue("h", "help", "Show help of task"),
-    }
-
-    identity = AflutterTaskIdentity(
-        "-parse-options",
-        "Parsing arguments",
-        _Dict.flatten(__options),
-        lambda: ParseOptions(),
-    )
+    def describe(self, args: Args) -> str:
+        return "Parsing arguments"
 
     def execute(self, args: Args) -> TaskResult:
-        # Fill options list with current tasks
-        from ..core.task.manager import TaskManager
+        from ..core.task import TaskManager
 
-        options: List[Option] = self.identity.options.copy()
-        for task in TaskManager._task_stack:
-            options.extend(task.options)
-        skip = (
-            not _Iterable.first_or_none(
-                options, lambda x: isinstance(x, OptionAll))
-            is None
-        )
+        long_options: Dict[Argument, Dict[Group, _Helper[LongOption]]] = {}
+        short_options: Dict[Argument, Dict[Group, _Helper[ShortOption]]] = {}
+        positional_options: Dict[Argument, Dict[Group, _Helper[PositionalOption]]] = {}
+        option_all: List[_Helper[OptionAll]] = []
 
-        if skip:
-            encoder = ArgsEncode(args)
-            for arg in sys_argv[2:]:
-                if arg == "--aflutter-stack-trace":
-                    Session.show_stacktrace = True
+        # Separate and identify options by type
+        for identity in TaskManager._task_stack.copy():
+            for option in identity.options:
+                if isinstance(option, OptionAll):
+                    option_all.append(_Helper(option, identity, OptionAll))
                     continue
-                elif arg in ("-h", "--help"):
-                    TaskManager._task_stack.clear()
-                    TaskManager.add(HelpStub(sys_argv[1]))
-                encoder.add(arg)
-            return TaskResult(args)
+                if isinstance(option, LongOption):
+                    _Helper(option, identity, LongOption).into(long_options)
+                if isinstance(option, ShortOption):
+                    _Helper(option, identity, ShortOption).into(short_options)
+                if isinstance(option, PositionalOption):
+                    _Helper(option, identity, PositionalOption).into(positional_options)
+            pass
 
-        short = ""
-        long: List[str] = []
-        for option in options:
-            if isinstance(option, LongOption):
-                if isinstance(option, OptionWithValue):
-                    long.append(option.long + "=")
+        input = sys_argv[2:]
+        has_param: List[_Helper] = []
+        maybe_has_param: Optional[_Helper[Union[LongOption, ShortOption]]] = None
+        position_count = 0
+        for argument in input:
+            for helper_all in option_all:
+                self.__append_argument(args, helper_all, argument)
+
+            # Last iteration require param
+            if len(has_param) > 0:
+                for helper_has_param in has_param:
+                    self.__append_argument(args, helper_has_param, argument)
+                has_param = []
+                continue
+
+            size = len(argument)
+            # Last iteration maybe require param
+            if not maybe_has_param is None:
+                if size > 1 and argument[0] == "-":
+                    if isinstance(maybe_has_param.option, ShortOption):
+                        self.__append_argument(
+                            args,
+                            _Helper(
+                                ShortOption(maybe_has_param.option.short, ""),
+                                maybe_has_param.group,
+                                ShortOption,
+                            ),
+                            None,
+                        )
+                    elif isinstance(maybe_has_param.option, LongOption):
+                        self.__append_argument(
+                            args,
+                            _Helper(
+                                LongOption(maybe_has_param.option.long, ""),
+                                maybe_has_param.group,
+                                LongOption,
+                            ),
+                            None,
+                        )
+                    maybe_has_param = None
                 else:
-                    long.append(option.long)
-                pass
-            if isinstance(option, ShortOption):
-                short += option.short
-                if isinstance(option, OptionWithValue):
-                    short += ":"
-                pass
-        try:
-            opts, positional = gnu_getopt(sys_argv[2:], short, long)
-        except GetoptError as error:
-            return TaskResult(args, error, success=False)
+                    self.__append_argument(args, maybe_has_param, argument)
+                    maybe_has_param = None
+                    continue
 
-        for opt, value in opts:
-            opt_strip = opt.lstrip("-")
-            found: Optional[Option] = None
-            if len(opt_strip) == 1:
-                found = _Iterable.first_or_none(
-                    options,
-                    lambda option: isinstance(option, ShortOption)
-                    and option.short == opt_strip,
-                )
-            else:
-                found = _Iterable.first_or_none(
-                    options,
-                    lambda option: isinstance(option, LongOption)
-                    and option.long == opt_strip,
-                )
-
-            if found is None:
-                # Argument is not parameter
-                args.add(Arg(opt, None))
-            elif isinstance(found, ShortOption):
-                # Using short argument
-                args.add(
-                    Arg(
-                        opt,
-                        value if isinstance(found, OptionWithValue) else None,
+            # Handle short option argument
+            if size == 2 and argument[0] == "-":
+                sub = argument[1:].lower()
+                if sub in short_options:
+                    for group, helper_short in short_options[sub].items():
+                        if helper_short.has_value:
+                            has_param.append(helper_short)
+                        else:
+                            self.__append_argument(args, helper_short, None)
+                    continue
+                else:
+                    raise OptionNotFound(
+                        "Unrecognized command line option {}".format(argument)
                     )
-                )
-            elif isinstance(found, LongOption):
-                args.add(
-                    Arg(
-                        "--" + found.long,
-                        value if isinstance(found, OptionWithValue) else None,
+
+            elif size >= 4 and argument[0] == "-" and argument[1] == "-":
+
+                split = argument[2:].lower().split(":")
+                split_len = len(split)
+                if split_len == 1:
+                    sub = split[0]
+                    group = None
+                elif split_len == 2:
+                    sub = split[1]
+                    group = split[0]
+                else:
+                    raise OptionInvalidFormat(
+                        "Invalid argument group structure for command line option {}".format(
+                            argument
+                        )
                     )
-                )
+                # Short argument with group
+                if len(sub) == 1:
+                    if sub in short_options:
+                        for group, helper_short in short_options[sub].items():
+                            if helper_short.has_value:
+                                has_param.append(helper_short)
+                            else:
+                                self.__append_argument(args, helper_short, None)
+                        continue
+                    elif not group is None:
+                        maybe_has_param = _Helper(
+                            _ShortOptionMaybeWithValue(sub, ""), group, ShortOption
+                        )
+                        continue
+                    else:
+                        raise OptionNotFound(
+                            "Unrecognized command line option {}".format(argument)
+                        )
+
+                # Long argument
+                if sub in long_options:
+                    if group is None:
+                        for grp, helper_long in long_options[sub].items():
+                            if helper_long.has_value:
+                                has_param.append(helper_long)
+                            else:
+                                self.__append_argument(args, helper_long, None)
+                        continue
+                    else:
+                        if group in long_options[sub]:
+                            helper_long = long_options[sub][group]
+                            if helper_long.has_value:
+                                has_param.append(helper_long)
+                            else:
+                                self.__append_argument(args, helper_long, None)
+                            continue
+                        else:
+                            # unregistered group
+                            maybe_has_param = _Helper(
+                                _LongOptionMaybeWithValue(sub, ""), group, LongOption
+                            )
+                            continue
+                elif not group is None:
+                    # unregistered option with group
+                    maybe_has_param = _Helper(
+                        _LongOptionMaybeWithValue(sub, ""), group, LongOption
+                    )
+                    continue
+                else:
+                    raise OptionNotFound(
+                        "Unrecognized command line option {}".format(argument)
+                    )
+
             else:
-                raise TypeError(
-                    "Unexpected instance of Option: {}".format(
-                        type(found).__name__)
-                )
-
-        i = 0
-        for value in positional:
-            args["-" + str(i)] = Arg(value, None)
-            i += 1
-
-        if "aflutter-stack-trace" in args:
-            Session.show_stacktrace = True
-            args.pop("aflutter-stack-trace")
-        if args.contains(self.__options["help"]):
-            TaskManager._task_stack.clear()
-            TaskManager.add(HelpStub(sys_argv[1]))
+                # Positional argument
+                pos = str(position_count)
+                if not pos in positional_options:
+                    raise OptionNotFound(
+                        'Unrecognized positional command line "{}"'.format(argument)
+                    )
+                for group, helper_positional in positional_options[pos].items():
+                    self.__append_argument(args, helper_positional, argument)
+            pass
 
         return TaskResult(args)
+
+    def __append_argument(self, args: Args, helper: _Helper, value: Optional[str]):
+        option: Option = helper.option
+        group: Group = helper.group
+        if helper.has_value and value is None:
+            raise OptionRequireValue(
+                "Command line {} requires value, but nothing found"
+            )
+        if isinstance(option, OptionAll):
+            assert not value is None
+            args.group_add_all(group, option, value)
+            return
+        args.group_add(group, option, value)
