+++ src/auto_flutter_dig/task/setup/setup.py
@@ -6,7 +6,6 @@ from ...model.task import *
 from ...task.identity import AflutterTaskIdentity
 from ..firebase import FirebaseCheck
 from ..flutter import FlutterCheck
-from ..options import ParseOptions
 from .edit import SetupEdit
 
 
@@ -19,7 +18,7 @@ class Setup(Task):
     )
 
     def require(self) -> List[TaskId]:
-        return [ParseOptions.identity.id, SetupEdit.identity.id]
+        return [SetupEdit.identity.id]
 
     def describe(self, args: Args) -> str:
         if args.contains(SetupEdit.option_show):
